IDVA6-1630 move csrf middleware into its own file
