+++ src/middleware/session.middleware.ts
@@ -1,9 +1,8 @@
 import { SessionMiddleware, SessionStore, CookieConfig } from "@companieshouse/node-session-handler";
 import Redis from "ioredis";
 import * as constants from "../lib/constants";
-import { CsrfProtectionMiddleware } from "@companieshouse/web-security-node";
 
-const sessionStore = new SessionStore(new Redis(`redis://${constants.CACHE_SERVER}`));
+export const sessionStore = new SessionStore(new Redis(`redis://${constants.CACHE_SERVER}`));
 
 const cookieConfig: CookieConfig = {
     cookieName: constants.COOKIE_NAME,
@@ -14,9 +13,3 @@ const cookieConfig: CookieConfig = {
 };
 
 export const sessionMiddleware = SessionMiddleware(cookieConfig, sessionStore, true);
-
-export const csrfProtectionMiddleware = CsrfProtectionMiddleware({
-    sessionStore,
-    enabled: true,
-    sessionCookieName: constants.COOKIE_NAME
-});
