+++ test/mocks/csrf.protection.middleware.mock.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from "express";
-import { csrfProtectionMiddleware } from "../../src/middleware/session.middleware";
+import { csrfProtectionMiddleware } from "../../src/middleware/csrf.protection.middleware";
 
-jest.mock("../../src/middleware/session.middleware");
+jest.mock("../../src/middleware/csrf.protection.middleware");
 
 const mockCsrfProtectionMiddleware = csrfProtectionMiddleware as jest.Mock;
